+++ src/Tests/Unit/Private/Invoke-XMLDataCheck.Tests.ps1
@@ -12,9 +12,7 @@ Import-Module $PathToManifest -Force
 #-------------------------------------------------------------------------
 
 InModuleScope 'PSGalleryExplorer' {
-    #-------------------------------------------------------------------------
-    $WarningPreference = "SilentlyContinue"
-    #-------------------------------------------------------------------------
+
     Describe 'Invoke-XMLDataCheck' -Tag Unit {
         BeforeAll {
             $WarningPreference = 'SilentlyContinue'
@@ -25,9 +23,12 @@ InModuleScope 'PSGalleryExplorer' {
             }
             function Expand-XMLDataSet {
             }
-            function Get-XMLDataSet {
+            function Get-RemoteFile {
+            }
+            function Remove-Item {
             }
         } #before_all
+
         Context 'ShouldProcess' {
             BeforeEach {
                 Mock -CommandName Invoke-XMLDataCheck -MockWith { } #endMock
@@ -63,7 +64,8 @@ InModuleScope 'PSGalleryExplorer' {
                 Invoke-XMLDataCheck -Force
                 Assert-MockCalled Invoke-XMLDataCheck -Scope It -Exactly -Times 1
             } #it
-        }
+        } #context_shouldprocess
+
         BeforeEach {
             Mock -CommandName Confirm-DataLocation -MockWith {
                 $true
@@ -74,29 +76,38 @@ InModuleScope 'PSGalleryExplorer' {
             Mock -CommandName Expand-XMLDataSet -MockWith {
                 $true
             } #endMock
-            Mock -CommandName Get-XMLDataSet -MockWith {
+            Mock -CommandName Get-RemoteFile -MockWith {
                 $true
             } #endMock
+            Mock -CommandName Test-Path -MockWith {
+                $true
+            } #endMock
+            Mock -CommandName Remove-Item -MockWith {}
         } #before_each
+
         Context 'Success' {
+
             It 'should return true if the data file is confirmed' {
                 Invoke-XMLDataCheck -Force | Should -BeExactly $true
             } #it
+
             It 'should return false if the data output dir cannot be confirmed' {
                 Mock -CommandName Confirm-DataLocation -MockWith {
                     $false
                 } #endMock
                 Invoke-XMLDataCheck -Force | Should -BeExactly $false
-            }
+            } #it
+
             It 'should return false if the data file is not confirmed and the file can not be downloaded' {
                 Mock -CommandName Confirm-XMLDataSet -MockWith {
                     $false
                 } #endMock
-                Mock -CommandName Get-XMLDataSet -MockWith {
+                Mock -CommandName Get-RemoteFile -MockWith {
                     $false
                 } #endMock
                 Invoke-XMLDataCheck -Force | Should -BeExactly $false
             } #it
+
             It 'should return false if the data file is not confirmed and the file is downloaded, but can not be expanded' {
                 Mock -CommandName Confirm-XMLDataSet -MockWith {
                     $false
@@ -106,12 +117,35 @@ InModuleScope 'PSGalleryExplorer' {
                 } #endMock
                 Invoke-XMLDataCheck -Force | Should -BeExactly $false
             } #it
+
             It 'should return true if the data file is not confirmed and the file is downloaded and expanded' {
                 Mock -CommandName Confirm-XMLDataSet -MockWith {
                     $false
                 } #endMock
                 Invoke-XMLDataCheck -Force | Should -BeExactly $true
             } #it
+
+            It 'should download 2 files if the data file is not confirmed' {
+                Mock -CommandName Confirm-XMLDataSet -MockWith {
+                    $false
+                } #endMock
+                Invoke-XMLDataCheck -Force
+                Should -Invoke Get-RemoteFile -Scope It -Exactly -Times 2
+            } #it
+
+            It 'should not attempt to remove the metadata file if it does not exist' {
+                Mock -CommandName Confirm-XMLDataSet -MockWith {
+                    $false
+                } #endMock
+                Mock -CommandName Test-Path -MockWith {
+                    $false
+                } #endMock
+                Invoke-XMLDataCheck -Force
+                Should -Invoke Remove-Item -Scope It -Exactly -Times 0
+            } #it
+
         } #context_Success
+
     } #describe_Invoke-XMLDataCheck
+
 } #inModule